The ErrorBoundary is super powerful and useful.  There is a default implementation that seems to be recommended now, so dropping our homegrown one for the standard one.
We can now render custom components when an error happens within an error boundary, and we can use error boundaries within sub-components as we now do in the reactFormBuilder which will catpure form rendering errors, and allow you to fix the error and retry.

The more global ErrorBoundary set in the "ContainerForExtensions" now users a the ErrorBoundaryFallack to render the error - which looks a little cleaner, and tries to offer a little more information about what went wrong.