Only return form related task data for human tasks (#2654)
If a user task is configured with a `variableName`, limit the task data
in the api response to only contain that key. This piggybacks on the
previous `variableName` work to not leak irrelevant task data.

I'm still auditing endpoints but this appears to be working and should
be safe to merge since it is only in play if a `variableName` is
configured.

<!-- This is an auto-generated comment: release notes by coderabbit.ai
-->

## Summary by CodeRabbit

* **Improvements**
* Enhanced task data handling with support for variable name extensions
in task configurations, enabling better organization and structure of
task-related information.

<sub>✏️ Tip: You can customize this high-level summary in your review
settings.</sub>

<!-- end of auto-generated comment: release notes by coderabbit.ai -->